requirements_pytorch.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.5.0
scikit-learn>=1.1.0
seaborn>=0.11.0
pillow>=9.0.0
